refactor(ios-keyboard): separate logo bar transport state from proprietary visuals
- move keyboard logo transport state, accessibility, and playback presentation inputs into KeyboardTransportDisplayState
- keep KeyboardLogoBarView focused on proprietary drawing, layout, and animation behavior only
- update the iOS codemap and engineering notes to document the visual-only licensing boundary
